move FeedForwardNetwork to own file
